コード整理
* glb/gltf ロード機能をオミット => GltfViewer を参照
* inner class SimpleViewer.Loaded を別ファイルに分離
* FileDialog を別フォルダに分離
