+++ Assets/VRM_Samples/SimpleViewer/FileDialog.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5b0aac758a2a5ab42aab2ccf5cf5480a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
