+++ Assets/VRM_Samples/SimpleViewer/FileDialog/FileDialogForWindows.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8a56b2b48ad6a1a4897d1de96cac39db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
