+++ Assets/VRM_Samples/SimpleViewer/FileDialog/FileUtil.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 952ae2eb9c0c59b4ba3ed6344bd867c4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
