+++ Assets/VRM_Samples/SimpleViewer/FileDialog/WebGLUtil.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b12500e13ead097488e89244d024ed2d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
