+++ Assets/VRM_Samples/SimpleViewer/Loaded.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 649d78dc96acf5b4f83d54cd1859e940
-timeCreated: 1524038001
-licenseType: Free
+guid: cd86f9044da608b4883f7bec5167a160
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
